feat(PresentProofV2): implemented presentation proposal
