+++ packages/core/src/agent/Agent.ts
@@ -19,6 +19,7 @@ import { ConnectionsModule } from '../modules/connections/ConnectionsModule'
 import { CredentialsModule } from '../modules/credentials/CredentialsModule'
 import { DiscoverFeaturesModule } from '../modules/discover-features'
 import { LedgerModule } from '../modules/ledger/LedgerModule'
+import { ProofsAPI } from '../modules/proofs/ProofsAPI'
 import { ProofsModule } from '../modules/proofs/ProofsModule'
 import { MediatorModule } from '../modules/routing/MediatorModule'
 import { RecipientModule } from '../modules/routing/RecipientModule'
@@ -46,7 +47,8 @@ export class Agent {
   public messageSubscription: Subscription
 
   public readonly connections!: ConnectionsModule
-  public readonly proofs!: ProofsModule
+  // public readonly proofs!: ProofsModule
+  public readonly proofs!: ProofsAPI
   public readonly basicMessages!: BasicMessagesModule
   public readonly ledger!: LedgerModule
   public readonly credentials!: CredentialsModule
@@ -96,7 +98,8 @@ export class Agent {
     // We set the modules in the constructor because that allows to set them as read-only
     this.connections = this.container.resolve(ConnectionsModule)
     this.credentials = this.container.resolve(CredentialsModule)
-    this.proofs = this.container.resolve(ProofsModule)
+    // this.proofs = this.container.resolve(ProofsModule)
+    this.proofs = this.container.resolve(ProofsAPI)
     this.mediator = this.container.resolve(MediatorModule)
     this.mediationRecipient = this.container.resolve(RecipientModule)
     this.basicMessages = this.container.resolve(BasicMessagesModule)
