+++ packages/core/src/modules/proofs/ProofProtocolVersion.ts
@@ -1,0 +1,4 @@
+export enum ProofProtocolVersion {
+  V1_0 = '1.0',
+  V2_0 = '2.0',
+}
