+++ packages/core/src/modules/proofs/v1/messages/index.ts
@@ -1,6 +1,5 @@
 export * from './ProposePresentationMessage'
 export * from './RequestPresentationMessage'
 export * from './PresentationMessage'
-export * from './PresentationPreview'
 export * from './PresentationAckMessage'
 export * from './PresentationProblemReportMessage'
