+++ packages/core/src/modules/proofs/v2/ProofRole.ts
@@ -1,0 +1,6 @@
+// ProofRole comes from present proof v2 roles
+// https://github.com/hyperledger/aries-rfcs/blob/main/features/0454-present-proof-v2/README.md#roles
+export enum ProofRole {
+  Verifier = 'verifier',
+  Prover = 'prover',
+}
